FileURLDownloadTask:  Check for truncated files
Check for the case of a server closing the socket before sending the entire
response body.  Also check for the unexpected case of a server sending more
bytes than indicated by the Content-Length headier field, which is probably
a sign of a server malfunction or a malicious attack, so it's best to throw
out the result and start over.

The total count of bytes read can exceed the content length if the response
body is compressed (as mobile network proxies are often wont to do), but we
opt out so that doesn't happen.