Link icon to the GitP Web site
The icon is offered as a link banner, so make it a link.

I would have linked to the page for the Order of the Stick if there were
one.  The table of contents page listing the newest comic titles doesn't
seem like the right target, so I simply used <http://www.giantitp.com/>.